move UI to Styleguide
