--- weg-li/UI/EditButtonStyle.swift
@@ -1,13 +1,0 @@
-// Created for weg-li in 2021.
-
-import SwiftUI
-
-struct EditButtonStyle: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .padding()
-            .foregroundColor(Color(.label))
-            .background(Color(.tertiarySystemFill))
-            .clipShape(RoundedRectangle(cornerRadius: 10))
-    }
-}
